+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.0 (December 19, 2021)
+
+### Release 1.1.0
+* New Features: 
+  - *Cursor::new_with_extras::<`Extras`>(*`slice`*);*
+  - *StrCursor::new_with_extras::<`Extras`>(*`str`*);*
+
+---
+
 ## 1.0.0 (December 19, 2021)
 
 ### Release 1.0.0